backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.29
pymysql==1.1.0
cryptography==42.0.5
pydantic==2.6.4
pydantic-settings==2.2.1

# ML
xgboost==2.0.3
scikit-learn==1.4.1.post1
tensorflow==2.16.1
joblib==1.3.2
numpy==1.26.4
pandas==2.2.1
Pillow==10.2.0

# API External
requests==2.31.0
httpx==0.27.0
groq==0.5.0

# Utils
python-dotenv==1.0.1
aiofiles==23.2.1